dataOps-service/requirements.txt
pandas
numpy
flask
snowflake-connector-python[pandas]
simple-salesforce
pycryptodomex
python-dotenv
cryptography
